Pull all signal-handling code to a common file.
