+++ cosmos/models/Workflow.py
@@ -3,13 +3,10 @@ Tools for defining, running and terminating Cosmos workflows.
 """
 
 import atexit
-import collections
 import datetime
 import os
 import re
-import signal
 import sys
-import threading
 import time
 import types
 
@@ -52,145 +49,6 @@ def _workflow_status_changed(ex):
     if ex.status == WorkflowStatus.successful:
         ex.successful = True
         ex.finished_on = datetime.datetime.now()
-
-
-class SignalWatcher(object):
-    """
-    Monitors the specified signals and sets an Event when one is caught.
-
-    Depending on its configuration, SGE can send a SIGUSR1, SIGUSR2, and/or
-    SIGXCPU before sending SIGSTOP/SIGKILL signals which cannot be caught.
-
-    According to ``man qsub`` (search for -notify), SGE can:
-
-    > send "warning" signals to a running job prior to sending the
-    > signals themselves. If a SIGSTOP is pending, the job will
-    > receive a SIGUSR1 several seconds before the SIGSTOP. If a
-    > SIGKILL is pending, the job will receive a SIGUSR2 several
-    > seconds before the SIGKILL.
-
-    acording to ``man queue_conf``, SGE additionally can:
-
-    > If s_cpu is exceeded, the job is sent a SIGXCPU signal which
-    > can be caught by the job.... If s_vmem is exceeded, the job
-    > is sent a SIGXCPU signal which can be caught by the job....
-    > If s_rss is exceeded, the job is sent a SIGXCPU signal which
-    > can be caught by the job....
-
-    Unfortunately, SIGUSR1 can mean two different things. As above, when -notify
-    is set, it means a SIGSTOP is imminent. However, and less frequently, it can
-    also mean a job is about to be killed. See, again, ``man queue_conf``:
-
-    > If h_rt is exceeded by a job running in the queue, it is aborted via the
-    > SIGKILL signal (see kill(1)). If s_rt is exceeded, the job is first
-    > "warned"  via the SIGUSR1 signal (which can be caught by the job) and
-    > finally aborted after the notification time defined in the queue
-    > configuration parameter notify (see above) has passed.
-
-    SIGUSR1 signals, then, are more likely than not to be a benign indication
-    that a job is about to be paused (since timeouts are far rarer than pauses).
-    However, when one is caught, we log that it may be a sign the job is about
-    to die.
-    """
-
-    # TODO Is it worth making signals configurable based on the caller's SGE
-    # TODO configuration? e.g. PRE_QDEL_SIGNAL=SIGUSR2, PRE_SUSP_SIGNAL=SIGUSR1,
-    # TODO PRE_TIME_EXC_SIGNAL=SIGUSR1, PRE_CPU_EXC_SIGNAL=SIGXCPU.
-
-    def __init__(self,
-                 workflow,
-                 lethal_signals=frozenset([signal.SIGINT, signal.SIGTERM, signal.SIGUSR2, signal.SIGXCPU]),
-                 benign_signals=frozenset([signal.SIGUSR1]),
-                 explanations={
-                     signal.SIGUSR1: 'SGE is about to send a SIGSTOP, or, if a time limit has been exceeded, a SIGKILL',
-                     signal.SIGUSR2: 'SGE is about to send a SIGKILL',
-                     signal.SIGXCPU: 'SGE is about to send a SIGKILL, because a cpu resource limit has been exceeded'}):
-
-        self.workflow = workflow
-        self.lethal_signals = lethal_signals
-        self.benign_signals = benign_signals
-        self.explanations = explanations
-
-        self._prev_handlers = dict()
-        self._signals_caught = collections.Counter()
-        self._signals_logged = collections.Counter()
-
-        self._logging_enabled = False
-        self._logging_event = None
-        self._logging_daemon = None
-
-    def __enter__(self):
-        self._logging_enabled = True
-        self._logging_event = threading.Event()
-        self._logging_daemon = threading.Thread(target=self.logging_daemon)
-        self._logging_daemon.daemon = True
-
-        self._logging_daemon.start()
-
-        for sig in self.lethal_signals | self.benign_signals:
-            self._cache_existing_handler(sig)
-            signal.signal(sig, self.signal_handler)
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        for sig, handler in self._prev_handlers.items():
-            signal.signal(sig, handler)
-
-        self._prev_handlers.clear()
-
-        self._logging_enabled = False
-        self._logging_event.set()
-        self._logging_daemon.join(timeout=1)
-
-        self.workflow.log.info('Caught/processed %d/%d signal(s) while running',
-                               sum(self._signals_caught.values()),
-                               sum(self._signals_logged.values()))
-
-    def signal_handler(self, signum, frame):    # pylint: disable=unused-argument
-        self._signals_caught[signum] += 1
-        self._logging_event.set()
-
-        if signum in self.lethal_signals:
-            self.workflow.terminate_when_safe = True
-
-    def _cache_existing_handler(self, sig):
-        prev_handler = signal.getsignal(sig)
-        if prev_handler not in (signal.SIG_DFL, signal.SIG_IGN, signal.default_int_handler):
-            raise RuntimeError("a signal handler already exists for signal %d (%s): %s" % (sig, self._explain(sig), prev_handler))
-        self._prev_handlers[sig] = prev_handler
-
-    def _explain(self, signum):
-        names = []
-        for k, v in signal.__dict__.iteritems():
-            if k.startswith('SIG') and v == signum:
-                names.append(k)
-        names.sort()
-
-        if signum in self.explanations:
-            return ': '.join((' or '.join(names), self.explanations[signum]))
-        else:
-            return ' or '.join(names)
-
-    def _log_signal_receipt(self, signal_counter):
-        for sig, cnt in signal_counter.items():
-            self.workflow.log.info('Caught signal %d %s(%s)',
-                                   sig, '%d times ' % cnt if cnt > 1 else '',
-                                   self._explain(sig))
-
-    def logging_daemon(self):
-        while self._logging_enabled:
-            self._logging_event.wait()
-            new_signals = self._signals_caught - self._signals_logged
-            self._logging_event.clear()
-
-            if new_signals:
-                self._log_signal_receipt(new_signals)
-                self._signals_logged += new_signals
-
-                if self.workflow.terminate_when_safe:
-                    self.workflow.log.info('Early-termination flag has been set')
-                else:
-                    self.workflow.log.debug('Ignoring benign signal(s)')
 
 
 class Workflow(Base):
@@ -623,45 +481,44 @@ def _run(workflow, session, task_queue):
     workflow.log.info('Executing TaskGraph')
     available_cores = True
 
-    with SignalWatcher(workflow):
-        while len(task_queue) > 0:
-            if available_cores:
-                _run_queued_and_ready_tasks(task_queue, workflow)
-                available_cores = False
+    while len(task_queue) > 0:
+        if available_cores:
+            _run_queued_and_ready_tasks(task_queue, workflow)
+            available_cores = False
 
-            for task in _process_finished_tasks(workflow.jobmanager):
-                if task.status == TaskStatus.failed and task.must_succeed:
+        for task in _process_finished_tasks(workflow.jobmanager):
+            if task.status == TaskStatus.failed and task.must_succeed:
 
-                    if workflow.info['fail_fast']:
-                        workflow.log.info('Exiting run loop at first Task failure')
-                        workflow.terminate(due_to_failure=True)
-                        return
+                if workflow.info['fail_fast']:
+                    workflow.log.info('Exiting run loop at first Task failure')
+                    workflow.terminate(due_to_failure=True)
+                    return
 
-                    # pop all descendents when a task fails; the rest of the graph can still execute
-                    remove_nodes = descendants(task_queue, task).union({task, })
-                    # graph_failed.add_edges(task_queue.subgraph(remove_nodes).edges())
+                # pop all descendents when a task fails; the rest of the graph can still execute
+                remove_nodes = descendants(task_queue, task).union({task, })
+                # graph_failed.add_edges(task_queue.subgraph(remove_nodes).edges())
 
-                    task_queue.remove_nodes_from(remove_nodes)
-                    workflow.status = WorkflowStatus.failed_but_running
-                    workflow.log.info('%s tasks left in the queue' % len(task_queue))
-                elif task.status == TaskStatus.successful:
-                    # just pop this task
-                    task_queue.remove_node(task)
-                elif task.status == TaskStatus.no_attempt:
-                    # the task must have failed, and is being reattempted
-                    pass
-                else:
-                    raise AssertionError('Unexpected finished task status %s for %s' % (task.status, task))
-                available_cores = True
+                task_queue.remove_nodes_from(remove_nodes)
+                workflow.status = WorkflowStatus.failed_but_running
+                workflow.log.info('%s tasks left in the queue' % len(task_queue))
+            elif task.status == TaskStatus.successful:
+                # just pop this task
+                task_queue.remove_node(task)
+            elif task.status == TaskStatus.no_attempt:
+                # the task must have failed, and is being reattempted
+                pass
+            else:
+                raise AssertionError('Unexpected finished task status %s for %s' % (task.status, task))
+            available_cores = True
 
-            # only commit Task changes after processing a batch of finished ones
-            session.commit()
+        # only commit Task changes after processing a batch of finished ones
+        session.commit()
 
-            time.sleep(.3)
-            if workflow.terminate_when_safe:
-                workflow.log.info('%s Stopping workflow due to signal(s)', workflow)
-                workflow.terminate(due_to_failure=False)
-                return
+        time.sleep(.3)
+        if workflow.terminate_when_safe:
+            workflow.log.info('%s Stopping workflow due to signal(s)', workflow)
+            workflow.terminate(due_to_failure=False)
+            return
 
 def _run_queued_and_ready_tasks(task_queue, workflow):
     max_cores = workflow.max_cores
